requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
flask-migrate
passlib
gunicorn
psycopg2
requests
rq
redis
Flask-RQ2
flask-uploads